requirements.txt
albumentations==1.3.1
coremltools==7.0
Flask==3.0.0
Flask_Mail==0.9.1
flask_wtf==1.2.1
matplotlib==3.5.3
numpy==1.23.5
onnx==1.15.0
onnx_graphsurgeon==0.0.1.dev5
onnxsim==0.4.35
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
pafy==0.5.5
pandas==2.0.3
Pillow==10.0.1
Pillow==10.1.0
Pillow==10.1.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
setuptools==68.2.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
tritonclient==2.38.0
wandb==0.15.12
Werkzeug==3.0.1
WTForms==3.1.0
